Implemented local storage and useTodoState hook
Added useTodoState.js. Changes made to TodoApp.js and TodoList.js.
