+++ src/components/TodoList.js
@@ -6,32 +6,34 @@ import Divider from "@material-ui/core/Divider";
 
 const TodoList = props => {
   const { todos, removeTodo, toggleTodo, editTodo } = props;
-  return (
-    <Paper
-      style={{
-        backgroundColor: "white"
-      }}
-      elevation={12}
-      width={75}
-    >
-      <List style={{ backgroundColor: "white" }}>
-        {todos.map((todo, i) => (
-          <Fragment>
-            <Todo
-              task={todo.task}
-              completed={todo.completed}
-              key={todo.id}
-              id={todo.id}
-              removeTodo={removeTodo}
-              toggleTodo={toggleTodo}
-              editTodo={editTodo}
-            />
-            {i < todos.length - 1 && <Divider />}
-          </Fragment>
-        ))}
-      </List>
-    </Paper>
-  );
+  if (todos.length)
+    return (
+      <Paper
+        style={{
+          backgroundColor: "white"
+        }}
+        elevation={12}
+        width={75}
+      >
+        <List style={{ backgroundColor: "white" }}>
+          {todos.map((todo, i) => (
+            <Fragment>
+              <Todo
+                task={todo.task}
+                completed={todo.completed}
+                key={todo.id}
+                id={todo.id}
+                removeTodo={removeTodo}
+                toggleTodo={toggleTodo}
+                editTodo={editTodo}
+              />
+              {i < todos.length - 1 && <Divider />}
+            </Fragment>
+          ))}
+        </List>
+      </Paper>
+    );
+  return null;
 };
 
 export default TodoList;
